Ex 2: Filter data sent to client
While it'd be better to get Prisma to do the filtering for us in this
case, I'm committing this version as it highlights a nice feature of
Remix's API end points.

You can filter out any data that your app doesn't require, cutting down
on the bandwidth and CPU required to process the data.