Ex 3: Satisfy TypeScript's complaints
Note that Remix won't ever let us get into the situation where the slug
will not be set inside the route component. We're only threatening to
throw that error to introduce type narrowing, so TypeScript doesn't
complain.

I do wonder whether it'd be better (clearer) to use a directive to tell
TypeScript to shut up.